rbd: thick image provisioning support
The command has the option specifying the thick mode for creating
new image, if applied.

If the option is specified, the command creates a thick image.
And it will take long time when a large thick image is created
because of writing actual data. The time depends on the size of image.

Signed-off-by: Hitoshi Kamei <hitoshi.kamei.xm@hitachi.com>
Cc: Mitsuo Hayasaka <mitsuo.hayasaka.hu@hitachi.com>